requirements.txt
azure-functions
gensim==3.8.1